Added has-children Class For Categories List
This is not the most beautiful implementation, but it shows how I solved it quick and dirty. Feel free to close the PR without merging and implement it in a nice way :) Finally, all links should have a `has-children` class if they have child content. This is currently only the case for the first level.